+++ include/Render.h
@@ -12,14 +12,14 @@ public:
 	Render(const Render& orig);
 	virtual ~Render();
 	void SetTexture(sf::Texture &tex);
-	void SetSprite(float u,float v,sf::IntRect s,float x, float y,float scale);
-	void Draw(sf::RenderWindow &window, const vector<float> &posBef,const vector<float> &posNow, float interpolation);
+	void SetSprite(double u,double v,sf::IntRect s,double x, double y,double scale);
+	void Draw(sf::RenderWindow &window, const vector<double> &posBef,const vector<double> &posNow, double interpolation);
 	sf::Sprite &GetSprite() { return sprite; }
-	vector<float> GetRenderPosition() const { return renderPos; }
+	vector<double> GetRenderPosition() const { return renderPos; }
 
 private:
 	sf::Sprite	sprite;
-	vector<float> renderPos;
+	vector<double> renderPos;
 };
 
 #endif // RENDER_H
